% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiTxDb.Hs.hg38.R
\docType{package}
\name{EpiTxDb.Hs.hg38}
\alias{EpiTxDb.Hs.hg38}
\alias{EpiTxDb.Hs.hg38.RMBase}
\alias{EpiTxDb.Hs.hg38.snoRNAdb}
\alias{EpiTxDb.Hs.hg38.tRNAdb}
\alias{chain.rRNA.hg19Tohg38}
\alias{chain.rRNA.hg38Tohg19}
\alias{snoRNA.targets.hg38}
\title{Annotation package for \code{EpiTxDb} objects}
\usage{
EpiTxDb.Hs.hg38.RMBase(version = "1")

EpiTxDb.Hs.hg38.snoRNAdb(version = "1")

EpiTxDb.Hs.hg38.tRNAdb(version = "1")

chain.rRNA.hg19Tohg38(version = "1")

chain.rRNA.hg38Tohg19(version = "1")

snoRNA.targets.hg38()
}
\arguments{
\item{version}{a \code{character} value defining a version. Versions
available: \code{"1"}.(default: \code{version = "1"})}
}
\value{
a \code{\link[EpiTxDb:EpiTxDb-class]{EpiTxDb}} object
}
\description{
This package loads one or more \code{EpiTxDb} objects. Such \code{EpiTxDb}
objects are an R interface to prefabricated databases contained by this
package.

The names of any objects exposed by this package indicate the origin and
resources exposed. So for example \code{EpiTxDb.Hs.hg38.snoRNAdb} would be a
\code{EpiTxDb} object of Homo sapiens data from snoRNAdb build based on the
hg38 genome.
}
\examples{
EpiTxDb.Hs.hg38.snoRNAdb()
}
\seealso{
\itemize{
\item{\code{\link[EpiTxDb:modifications]{modifications}}}
\item{\code{\link[EpiTxDb:modifications]{modificationsBy}}}
\item{\code{\link[EpiTxDb:modifications]{modifiedSeqsByTranscript}}}
}
}
\author{
Felix G M Ernst [aut]
}
