% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxModDb.Hsapiens.hg38.R
\docType{package}
\name{TxModDb.Hsapiens.hg38}
\alias{TxModDb.Hsapiens.hg38}
\title{Annotation package for TxModDb object(s)}
\description{
This package loads one or more TxModDb objects. Such TxModDb objects are an
R interface to prefabricated databases contained by this package.

The names of any objects exposed by this package indicate the origin and
resources exposed.  So for example TxModDb.Hsapiens.hg38.snoRNAdb would be a
TxModDb object of Homo sapiens data from snoRNAdb build based on the hg38
build.
}
\note{

}
\examples{
library(TxModDb.Hsapiens.hg38)
TxModDb.Hsapiens.hg38.snoRNAdb
}
\seealso{
\itemize{
\item{\code{\link[TxModDb:modifications]{modifications}}}
\item{\code{\link[TxModDb:modifications]{reactions}}}
\item{\code{\link[TxModDb:modifications]{specifies}}}
\item{\code{\link[TxModDb:modifications]{modificationsByTranscript}}}
\item{\code{\link[TxModDb:modifications]{modifiedSeqsByTranscript}}}
}
}
\author{
Felix G M Ernst [aut]
}
